slurmstepd - Fix _handle_completion for REQUEST_STEP_COMPLETE retries
If child slurmstepd get a timeout when sending a REQUEST_STEP_COMPLETE
to its parent it will retry it (_one_step_complete_msg()).
Those messages are handled on _handle_completion(), where it does the
jobacctinfo_aggregate().
This commit detects if the message was already processed to avoid errors on
the aggregated jobacctinto.

Bug 10723